% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{String or integer vector of year to be read. If data for 
one or more supplied years does not exist, execution will stop.}

\item{year}{Year as string or integer.}
}
\value{
A state map with all year accidents.
}
\description{
Read files by year references and map out state data for selected state.
}
\examples{
fars_map_state(state.num = 1, year = 2013)

}
