% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS csv data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the FARS csv data file. If file is not a csv file, 
an error will occur. If file does not exists, execution will stop.}
}
\value{
Content of file as a tibble.
}
\description{
Read a FARS csv data file and return the content as a tibble.
}
\examples{
my_data <- fars_read(filename = "path//to//my//csv//file.csv")

}
