% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple FARS data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{String vector or integer vector of year to be read. If data for 
one or more supplied years does not exist, execution will stop.}
}
\value{
List of tibbles.
}
\description{
Apply function (\code{fars_read}) to a set of files by year references.
}
\examples{
\dontrun{
my_data <- fars_read_years(years = c(2013, 2014, 2015))
}

}
