% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize multiple FARS data files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{String vector or integer vector of year to be read. If data for 
one or more supplied years does not exist, execution will stop.}
}
\value{
List of tibbles.
}
\description{
Read and summarize FARS files by year references (group by years).
}
\examples{
\dontrun{
my_data <- fars_summarize_years(years = c(2013, 2014, 2015))
}

}
