% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year as string or integer.}
}
\value{
A FARS data file name on the form  accident_yyyy.csv.bz2
}
\description{
Construct file name from year.
}
\examples{
\dontrun{
make_filename(year=2013)
make_filename(year="2014")
}
}
